Use gint instead of gfloat for indentation level.
2009-03-05  Ignacio Casal Quinteiro  <nacho.resa@gmail.com>

        * gtksourceview/simple-indenter.c:
        Use gint instead of gfloat for indentation level.

        * gtksourceview/gtksourceview.c:
        If we have any indentation when we press \n delete it      
        before insert the new indentation.


svn path=/branches/indenters/; revision=2202

+++ gtksourceview/simple-indenter.c
@@ -22,7 +22,7 @@
 
 #include "simple-indenter.h"
 #include "gtksourceindenter.h"
-#include "gtksourceview-utils.h"
+#include "gtksourceindenter-utils.h"
 
 #define SIMPLE_INDENTER_GET_PRIVATE(object)(G_TYPE_INSTANCE_GET_PRIVATE((object),\
 					    SIMPLE_TYPE_INDENTER, SimpleIndenterPrivate))
@@ -33,7 +33,7 @@ struct _SimpleIndenterPrivate
 
 G_DEFINE_TYPE (SimpleIndenter, simple_indenter, GTK_TYPE_SOURCE_INDENTER)
 
-static gfloat
+static gint
 simple_indenter_get_indentation_level (GtkSourceIndenter *indenter,
 				       GtkTextView *view,
 				       GtkTextIter *iter,
